Move RegionData to own file and motivate use
